requirements.txt
boto3==1.26.63
botocore==1.29.63
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
Flask==2.2.2
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.24.1
Pillow==9.4.0
proglog==0.1.10
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.0
